Add files via upload
What is Glassmorphism??
Glassmorphism UI is a card-based interface used for designing the user interface of websites, smartphones, web apps, mobile apps, etc.

How to implement glassmorphism??
1. Transparency
A blur behind the object or element keeps it from blending with the main background. This blurriness is not only essential to maintain readability and accessibility, but it also provides a slick frosted glass look.

2. Dark Background color.
Remember to have a dark background so that the transparency can differ from it.

3. Hierarchy.
Glassmorphism includes a shadow effect that gives the element some elevation and makes it appear floating above the surface. Therefore, the user will be able to establish a sense of depth and hierarchy while interacting with the UI.

In addition, it may have some shine on its edges to support its glass-like appearance.

In the case of stacked or nested elements, the closest object should have a more opaque background with a higher blur. On the other hand, the one at the farthest point should have more fill transparency and low background blurring.